+++ adminutils/actions.py
@@ -1,6 +1,8 @@
 import functools
 
 from django import http
+from django.db import transaction
+from django.shortcuts import render, redirect
 
 from django_object_actions import takes_instance_or_queryset
 
@@ -14,8 +16,10 @@ class MethodsRequiredDecorator(object):
         def wrapper(instance, request, *args, **kwargs):
             if request.method not in self.methods:
                 return http.HttpResponseBadRequest(
-                    '{} not allowed'.format(request.method))
+                    "{} not allowed".format(request.method)
+                )
             return func(instance, request, *args, **kwargs)
+
         return wrapper
 
 
@@ -28,4 +32,44 @@ def queryset_action(func):
 
 
 def object_action(func):
-    return require_method('POST')(func)
+    return require_method("POST")(func)
+
+
+def form_processing_action(
+    form_class, template_name="admin/generic_form.html", action_label=None
+):
+    def processor(func):
+        def view(self, request, queryset):
+            opts = self.model._meta
+            tool = getattr(self, request.resolver_match.kwargs["tool"])
+            tool_label = getattr(tool, "label")
+
+            if request.method == "POST":
+                form = form_class(request.POST, request.FILES)
+                if form.is_valid():
+                    with transaction.atomic():
+                        resp = func(self, request, form)
+                    if resp is None:
+                        resp = redirect(
+                            "admin:%s_%s_changelist"
+                            % (opts.app_label, opts.model_name)
+                        )
+                    return resp
+            else:
+                form = form_class()
+
+            return render(
+                request,
+                template_name,
+                {
+                    "form": form,
+                    "opts": opts,
+                    "app_label": opts.app_label,
+                    "tool_label": tool_label,
+                    "action_label": action_label,
+                },
+            )
+
+        return view
+
+    return processor