+++ HISTORY.rst
@@ -2,6 +2,14 @@
 History
 =======
 
+0.0.5 - 2020-03-16
+==================
+
+* Added a generic form processing action
+* Add an option to not use forms to submit actions
+* Compatibility with Django 2
+
+
 0.0.4 - 2017-11-23
 ==================